fixed following bug and moved logic to lib
